refactor: bet 패키지 분리
배팅 금액과 관련한 model들 이동
