+++ src/main/java/blackjack/model/Game.java
@@ -1,5 +1,7 @@
 package blackjack.model;
 
+import blackjack.model.bet.Bet;
+import blackjack.model.bet.Profits;
 import blackjack.model.player.Dealer;
 import blackjack.model.player.Entry;
 import blackjack.model.player.Players;
