+++ src/main/java/blackjack/model/player/Entries.java
@@ -1,8 +1,8 @@
 package blackjack.model.player;
 
-import blackjack.model.Bet;
-import blackjack.model.Profits;
-import blackjack.model.Result;
+import blackjack.model.bet.Bet;
+import blackjack.model.bet.Profits;
+import blackjack.model.bet.Result;
 import blackjack.model.trumpcard.TrumpCard;
 import java.util.List;
 import java.util.function.Supplier;
