+++ src/test/java/blackjack/model/bet/BetTest.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.bet;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
